+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added funtion `get_parameters()` to `BondGraph`.
+
 ## [0.2.0] 2023-04-30
 ### Changed
 - Breaking: Elements have been moved from `bondgraph.elements.basic` to `bondgraph.elements`.
 
 ### Removed
 - Custom elements are no longer included in library.
-- Removed obsolete field `parameters` from BondGraph. 
+- Removed obsolete field `parameters` from `BondGraph`. 
 
 ## [0.1.1] 2023-04-30
 ### Changed